% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredlmParam-class.R
\docType{class}
\name{PredlmParam-class}
\alias{PredlmParam-class}
\title{S4 class for the parameters for prediction based on a linear model}
\description{
Definition of the S4 class named \code{PredlmParam} for the parameters to make
predictions in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{RawData}}{\linkS4class{StQ} object with the raw data.}

\item{\code{EdData}}{\linkS4class{StQ} object with the edited data.}

\item{\code{VarNames}}{Character vector with the names of the variables whose probability errors are to
be computed.}

\item{\code{DomainNames}}{Character vector with the names of the variables determining the domains of the
population within which predictions are to be calculated separately.}

\item{\code{Imputation}}{\linkS4class{ImputationParam} object with the parameters to imputed missing
values during the computation of predicted values and their standard deviations (predition error
standard deviations).}
}}
\examples{
# An empty PredlmParam object:
new(Class = 'PredlmParam')

\dontrun{
ImpParam <- new(Class = 'PredlmParam',
                VarNames = 'CifraNeg_13.___',
                DomainNames =  'Tame_05._4.')
ErrorProbMLEParam <- new(Class = 'PredlmParam',
                         RawData = FD.StQList,
                         EdData = FG.StQList,
                         VarNames = 'CifraNeg_13.___',
                         Imputation = ImpParam)


}

}

